+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased](https://github.com/tbhb/typing-graph/compare/v0.3.0...HEAD)
 
+### Removed
+
+- `DiscriminatedUnionNode` class and `is_discriminated_union_node()` type guard
+- `TypeEdgeKind.VARIANT` edge kind
+
 ## [0.3.0](https://github.com/tbhb/typing-graph/compare/v0.2.0...v0.3.0) - 2025-12-10
 
 ### Added